third_party/WebKit/LayoutTests/fast/spatial-navigation/snav-textarea.html

<!DOCTYPE html>
  <head>
    <script src="../../resources/js-test.js"></script>
    <script src="resources/spatial-navigation-utils.js"></script>
  </head>
  <body id="some-content" xmlns="http://www.w3.org/1999/xhtml">
    <p id="description"></p>
    <table style="text-align: left; width: 100%; margin-left: auto; margin-right: auto;" border="1" cellpadding="2" cellspacing="1">
      <tbody>
        <tr>
          <td style="vertical-align: top; text-align: center;"></td>
          <td style="vertical-align: top; text-align: center;"><a id="2" href="a">2</a></td>
          <td style="vertical-align: top; text-align: center;"></td>
        </tr>
        <tr>
          <td style="vertical-align: top; text-align: center;"><a id="4" href="a">4</a></td>
          <td style="vertical-align: top; text-align: center;"><textarea id="start" rows="5" cols="2">abc d</textarea>
          <td style="vertical-align: top; text-align: center;"><a id="6" href="a">6</a></td>
        </tr>
        <tr>
          <td style="vertical-align: top; text-align: center;"></td>
          <td style="vertical-align: top; text-align: center;"><a id="8" href="a">8</a></td>
          <td style="vertical-align: top; text-align: center;"></td>
        </tr>
      </tbody>
    </table>
    <div id="console"></div>
    <script type="application/javascript">
    description('This test ensures the correctness of Spatial Navigation (SNav) algorithm over textarea.<br>\
                * Pre-conditions:<br>\
                1) DRT support for SNav enable/disable.<br>\
                * Navigation steps:<br>\
                1) Loads this page, focus goes to "start" automatically.<br>\
                2) Focus moves away from textarea in 4 different directions to neighbor nodes and back.<br>');

    var resultMap = [
      ["Down", "start"],
      ["Down", "start"],
      ["Down", "8"],
      ["Up", "start"], // The caret is at the end of TEXTAREA.
      ["Up", "start"], // The caret is at the beginning of TEXTAREA.
      ["Up", "2"],
      ["Down", "start"],
      ["Right", "start"],
      ["Right", "start"],
      ["Right", "start"],
      ["Right", "start"],
      ["Right", "start"], // The caret is at the end of TEXTAREA.
      ["Right", "6"],
      ["Left", "start"], // The caret is at the end of TEXTAREA.
      ["Left", "start"],
      ["Left", "start"],
      ["Left", "start"],
      ["Left", "start"],
      ["Left", "start"], // The caret is at the beginning of TEXTAREA.
      ["Left", "4"],
      ["Right", "start"],
      ["DONE", "DONE"]
    ];

    if (window.testRunner) {
      testRunner.dumpAsText();
      testRunner.overridePreference("WebKitTabToLinksPreferenceKey", 1);
      testRunner.overridePreference("WebKitSpatialNavigationEnabled", 1);
      testRunner.waitUntilDone();
    }

    function runTest()
    {
      var textarea = document.getElementById("start");
      textarea.value = "abc\nd";

      // starting the test itself: get to a known place.
      textarea.focus();
      textarea.setSelectionRange(0, 0);

      initTest(resultMap, testCompleted);
    }

    function testCompleted()
    {
      if (!window.testRunner)
        return;

      document.getElementById('start').value = "";
      var text = 'A text containing a space\nand a new line';
      for (var i = 0; i < text.length; ++i) {
        var key = text.charAt(i);
        if (key == '\n')
          key = 'Enter';
        eventSender.keyDown(key);
      }
      shouldBeEqualToString("document.getElementById('start').value", text);

      testRunner.notifyDone();
    }

    window.onload = runTest;

    </script>
  </body>
</html>
